Scaffolded PUT/DELETE schemas for /api/resume/[candidateId].
- Added resume PUT/DELETE route handlers (PUT/replace stubbed).
- Moved resume/candidate schema contracts.
- Replaced candidate DELETE endpoint with resume DELETE.
