+++ src/app/api/resume/[candidateId]/route.ts
@@ -1,7 +1,18 @@
 // Dependencies.
 import { type NextRequest, NextResponse } from "next/server"
-import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
-import { findResumeByCandidateId } from "@/lib/db/resume/sql"
+import { authorizeApiToken } from "@/lib/api/auth"
+import { resumePutSchema } from "@/lib/api/schemas/resume/contract"
+import {
+	parseJson,
+	validateDataFound,
+	validateRequestBodyAgainstSchema,
+	validateUuidFormat,
+} from "@/lib/api/validate"
+import {
+	deleteResumeByCandidateId,
+	findResumeByCandidateId,
+	replaceResumeByCandidateId,
+} from "@/lib/db/resume/sql"
 
 //
 // GET /api/resume/[candidateId].
@@ -27,4 +38,85 @@ export async function GET(
 	if (resumeValidationResponse) return resumeValidationResponse
 
 	return NextResponse.json({ resume }, { status: 200 })
+}
+
+//
+// PUT /api/resume/[candidateId].
+//
+export async function PUT(
+	request: NextRequest,
+	{ params }: { params: Promise<{ candidateId: string }> },
+) {
+	// Candidate ID.
+	const { candidateId } = await params
+
+	// Validate the candidate ID is a valid UUID.
+	const uuidFormatValidationResponse = validateUuidFormat(candidateId)
+	if (uuidFormatValidationResponse) return uuidFormatValidationResponse
+
+	// Authorize the API token.
+	const authorizationResponse = await authorizeApiToken(request, {
+		candidateId,
+		scopeRequirement: "resume:write",
+	})
+	if (authorizationResponse) return authorizationResponse
+
+	// Parse the request body JSON.
+	const requestBodyOrResponse = await parseJson(request)
+	if (requestBodyOrResponse instanceof NextResponse)
+		return requestBodyOrResponse
+
+	// Request body.
+	const requestBody = requestBodyOrResponse
+
+	// Validate the request body against the schema.
+	const resumePutOrResponse = validateRequestBodyAgainstSchema(
+		requestBody,
+		resumePutSchema,
+	)
+	if (resumePutOrResponse instanceof NextResponse) return resumePutOrResponse
+
+	// Replaced resume.
+	const replacedResume = await replaceResumeByCandidateId(candidateId)
+
+	// Validate the resume found.
+	const resumeValidationResponse = validateDataFound(replacedResume, "resume", {
+		candidateId,
+	})
+	if (resumeValidationResponse) return resumeValidationResponse
+
+	return NextResponse.json({ resume: replacedResume }, { status: 200 })
+}
+
+//
+// DELETE /api/resume/[candidateId].
+//
+export async function DELETE(
+	request: NextRequest,
+	{ params }: { params: Promise<{ candidateId: string }> },
+) {
+	// Candidate ID.
+	const { candidateId } = await params
+
+	// Validate the candidate ID is a valid UUID.
+	const uuidFormatValidationResponse = validateUuidFormat(candidateId)
+	if (uuidFormatValidationResponse) return uuidFormatValidationResponse
+
+	// Authorize the API token.
+	const authorizationResponse = await authorizeApiToken(request, {
+		candidateId,
+		scopeRequirement: "resume:write",
+	})
+	if (authorizationResponse) return authorizationResponse
+
+	// Deleted resume.
+	const deletedResume = await deleteResumeByCandidateId(candidateId)
+
+	// Validate the resume found.
+	const resumeValidationResponse = validateDataFound(deletedResume, "resume", {
+		candidateId,
+	})
+	if (resumeValidationResponse) return resumeValidationResponse
+
+	return new NextResponse(null, { status: 204 })
 }
